refactor: update components to use existingDatasets prop and consolidate name generation imports
