+++ DashAI/front/src/components/notebooks/datasetCreation/DataloaderConfiguration.jsx
@@ -1,22 +1,32 @@
 import { DialogContentText, Paper, Stack } from "@mui/material";
 import PropTypes from "prop-types";
-import React, { useState } from "react";
+import { useState, useMemo } from "react";
 import FormSchema from "../../shared/FormSchema";
 import FormSchemaLayout from "../../shared/FormSchemaLayout";
+import { generateDatasetName } from "../../../utils/nameGenerator";
+
 /**
  * This component is a form to configure a dataloader
  * @param {string} dataloader - The dataloader to configure
  * @param {function} onSubmit - The function to submit the form
  * @param {object} formSubmitRef - The reference to the form submit function
  * @param {function} setError - The function to set the error state
+ * @param {array} existingDatasets - Array of existing datasets to avoid name conflicts
  */
 function DataloaderConfiguration({
   dataloader,
   onSubmit,
   formSubmitRef,
   setError,
+  existingDatasets = [],
 }) {
   const [splitError, setSplitError] = useState(false);
+
+  const { defaultName } = useMemo(
+    () => generateDatasetName(existingDatasets),
+    [existingDatasets],
+  );
+
   const handleSubmitButtonClick = (values) => {
     onSubmit(values);
     setError(false);
@@ -27,7 +37,6 @@ function DataloaderConfiguration({
     <Paper sx={{ p: 4, height: "100%" }} borderRadius={2}>
       <Stack spacing={3}>
         {/* Form title */}
-
         <DialogContentText sx={{ alignSelf: "center" }}>
           {dataloader} configuration
         </DialogContentText>
@@ -36,22 +45,23 @@ function DataloaderConfiguration({
           <FormSchema
             autoSave
             model={dataloader}
-            onFormSubmit={(values) => {
-              handleSubmitButtonClick(values);
-            }}
+            onFormSubmit={handleSubmitButtonClick}
             formSubmitRef={formSubmitRef}
             setError={setError}
+            initialValues={{ name: defaultName }}
           />
         </FormSchemaLayout>
       </Stack>
     </Paper>
   );
 }
+
 DataloaderConfiguration.propTypes = {
   dataloader: PropTypes.string.isRequired,
   onSubmit: PropTypes.func.isRequired,
   formSubmitRef: PropTypes.shape({ current: PropTypes.any }),
   setError: PropTypes.func,
+  existingDatasets: PropTypes.array,
 };
 
 export default DataloaderConfiguration;
